+++ app/Jobs/Issues/SendEmailToFollowersAboutReminder.php
@@ -43,21 +43,36 @@ class SendEmailToFollowersAboutReminder implements ShouldQueue
         if (!is_null($this->issue->timeClosed)) {
             return;
         }
-        if ($this->typeOfReminder == 'paused') {
-            if (!$this->issue->paused) {
-                return null;
-            }
+
+        switch ($this->typeOfReminder) {
+            case 'paused':
+                if (!$this->issue->paused) {
+                    return null;
+                }
+                $delayDays = setting('days_reminder_paused_issue');
+                break;
+            case 'waitingForInternal':
+                if (!$this->issue->waitingForInternal) {
+                    return null;
+                }
+                $delayDays = setting('days_reminder_waiting_for_internal');
+                break;
+            case 'waitingForCustomer':
+                if (!$this->issue->waitingForCustomer) {
+                    return null;
+                }
+                $delayDays = setting('days_reminder_waiting_for_customer');
+                break;
+
+            default:
+                $delayDays = setting('days_reminder_waiting_for_comment');
+                if ($this->issue->latestComment->updated_at->addDays($delayDays) > now()) {
+                    return null;
+                }
+                break;
         }
-        if ($this->typeOfReminder == 'waitingForInternal') {
-            if (!$this->issue->waitingForInternal) {
-                return null;
-            }
-        }
-        if ($this->typeOfReminder == 'waitingForCustomer') {
-            if (!$this->issue->waitingForCustomer) {
-                return null;
-            }
-        }
+        // $delay = nextWorkingHour(now()->addDays($delayDays));
+        // CreateNewReminder::dispatch($this->issue, $this->typeOfReminder)->delay($delay);
         Mail::to($this->email)->send(new MailToFollowersAboutReminder($this->issue, $this->typeOfReminder));
     }
 }